% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_git_wd.R
\name{install_git_wd}
\alias{install_git_wd}
\title{wrapper to install git from network drives}
\usage{
install_git_wd(pdata_path, ...)
}
\arguments{
\item{pdata_path}{string path to network drive}

\item{...}{arguments for remotes::install_git}
}
\description{
wrapper to install git from network drives
}
