% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_reproducibility_chunk.R
\name{make_reproducibility_chunk}
\alias{make_reproducibility_chunk}
\title{Creates reproducibility tables automatically
just call in chunk with results = "asis"}
\usage{
make_reproducibility_chunk()
}
\value{
invisible
}
\description{
Creates reproducibility tables automatically
just call in chunk with results = "asis"
}
\examples{
make_reproducibility_chunk()
}
