% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_rmd.R
\name{source_rmd}
\alias{source_rmd}
\title{Source an R Markdown file}
\usage{
source_rmd(file, envir = .GlobalEnv, chdir = FALSE)
}
\arguments{
\item{file}{Path to the `.Rmd` file.}

\item{envir}{Environment in which to evaluate the code.
Defaults to the global environment (`.GlobalEnv`).}

\item{chdir}{Logical; if `TRUE`, the working directory is temporarily
changed to the location of `file` before sourcing, and reset afterwards.
Defaults to `FALSE`.}
}
\value{
Invisibly returns the result of [base::source()].
  Side effect: evaluated objects are created in \code{envir}.
}
\description{
Extracts and executes all R code chunks from an R Markdown file,
similar to [base::source()] but for `.Rmd`. Only chunks with `eval = TRUE`
are included (the same ones that would run when knitting).
}
\details{
By default, code is evaluated in the global environment (`.GlobalEnv`),
which mimics the most common interactive use of [base::source()]. You can
override this with the `envir` argument to isolate execution.
}
\examples{
\dontrun{
# Interactive use: objects go into the global environment
source_rmd("analysis.Rmd")

# Run into a custom environment
e <- new.env()
source_rmd("analysis.Rmd", envir = e)
ls(e)

# Respect relative paths inside the Rmd
source_rmd("analysis.Rmd", chdir = TRUE)
}

}
